feat: add `info` subcommand to show computed ranges & labels
`chart-review info` will now print some useful project stats to the
console. Namely, a list of annotators and their note count & ranges,
as well as a list of the final computed labels.

We can augment this over time or add flags to it. But this felt like
a reasonable start.