feat: add chart range output to `accuracy`
When running the accuracy command, print a nicer header giving some
basic info like: who is truth and who is merely annotator? How many
charts are in their overlap and what are the ID ranges?

This should give some peace of mind that the stats are covering what
they should be covering.

When passing in --save, we print the header to the console, but not
the written file, which is still just a csv/json of the stats.

Also:
- If a project didn't specify labels, the docs say we use all found
  labels. And we were... but only when scoring the "all labels" line
  item. We didn't score each found label on its own line. Now that's
  been fixed and we do score each found label separately.
- If an unknown annotator name is passed to the accuracy command, we
  now give a nice understandable error.