feat(info): add --ids command to print ID mappings
Top line feature is the addition of `info --ids` which will print all
the ID mappings (chart ID -> real FHIR ID -> anon FHIR ID).

It currently does this in CSV format to the console (which can be
redirected to a file as desired). I couldn't figure out a more pleasant
output format for console reading that would work for 50+ charts.
So for now, it's just a big techy CSV dump to the console.

Also:
- Refactor commands to isolate more of their logic into their modules,
  away from the topline cli.py which is now just a router to a command.
- Add missing help text to "accuracy --save"
- Print the set of ignored IDs (if any) in normal "info" output.