Allow no config at all
The only required field was `annotators` and that can have a
reasonable (if not entirely pleasant) default of just using the
Label Studio ID as the annotator name.

This way, it's one less friction point to starting to use
chart-review. You can always add nice names later.

While here, update the license definition and ruff config.